requirements.txt
hdx-python-api==3.2.5
